fix linux build

+++ arch/riscv/riscv-disas.h
@@ -513,8 +513,8 @@ typedef struct {
 
 /* functions */
 
-size_t inst_length(rv_inst inst);
-void inst_fetch(uint8_t *data, rv_inst *instp, size_t *length);
+int inst_length(rv_inst inst);
+void inst_fetch(uint8_t *data, rv_inst *instp, int *length);
 void disasm_inst(char *buf, size_t buflen, rv_isa isa, uint64_t pc, rv_inst inst);
 char * decodeRISCV(unsigned long int address, char *outbuf, int *lendis);
